Move execution stuff into its own context
