+++ src/pages/main.tsx
@@ -14,6 +14,7 @@ import NodeSelector from '../components/NodeSelectorPanel/NodeSelectorPanel';
 import ReactFlowBox from '../components/ReactFlowBox';
 import { BackendNodesResponse } from '../helpers/Backend';
 import { AlertBoxContext, AlertType } from '../helpers/contexts/AlertBoxContext';
+import { ExecutionProvider } from '../helpers/contexts/ExecutionContext';
 import { GlobalProvider } from '../helpers/contexts/GlobalNodeState';
 import { SettingsProvider } from '../helpers/contexts/SettingsContext';
 import { ipcRenderer } from '../helpers/safeIpc';
@@ -105,7 +106,9 @@ const Main = ({ port }: MainProps) => {
                         overflow="hidden"
                         p={2}
                     >
-                        <Header port={port} />
+                        <ExecutionProvider>
+                            <Header />
+                        </ExecutionProvider>
                         <HStack
                             as={Split}
                             defaultSplitterColors={{
